refactor: firebase 설정을 src/config/로 이동 및 import 업데이트
- fbase.js → src/config/firebase.ts로 이름 변경
- ReactoronConfig.ts → src/config/reactotron.ts로 이동
- firebase/ → src/firebase/로 이동
- 모든 32개 파일의 import를 @/config/firebase로 업데이트
- App.tsx의 Reactotron import 경로 업데이트
