+++ src/firebase/core/firestoreService.ts
@@ -1,4 +1,4 @@
-import { db } from '@/fbase';
+import { db } from '@/config/firebase';
 import {
 	DocumentData,
 	Query,
