+++ src/firebase/services/blockService.ts
@@ -1,4 +1,4 @@
-import { db } from '@/fbase';
+import { db } from '@/config/firebase';
 import { deleteDoc, doc, setDoc, Timestamp } from 'firebase/firestore';
 import firestoreRequest from '../core/firebaseInterceptor';
